change callback from clojure to hook method
